design: misc visual updates
