+++ src/components/DragPanel/styles.scss
@@ -3,7 +3,6 @@
 .dragPanel {
   @include row(center, center);
 
-  background-color: #f2f4f6;
   transition: all 0.25s ease-out;
   height: 100%;
 
@@ -11,11 +10,11 @@
   user-select: none;
 
   &:hover {
-    background-color: #edf1f3;
+    background-color: darken(#f6f6f6, 3);
   }
 
   &:active {
-    background-color: #e7eaec;
+    background-color: darken(#f6f6f6, 5);
     cursor: grabbing;
   }
 }
