+++ src/components/Map/NamespaceBackplate.tsx
@@ -27,7 +27,7 @@ export const NamespaceBackplate = memo(function NamespaceBackplate(
       transform={`translate(${props.xywh.x}, ${props.xywh.y})`}
     >
       <rect width={xywh.w} height={xywh.h} rx={10} ry={10} />
-      <text x={64} y={64}>
+      <text x={50} y={64}>
         {props.namespace}
       </text>
     </g>
