+++ src/components/Map/styles.scss
@@ -1,7 +1,7 @@
 .wrapper {
   display: block;
   width: 100%;
-  height: calc(100vh - var(--top-bar-height));
+  height: 100vh;
   cursor: grab;
 
   &:active {
@@ -12,14 +12,14 @@
 .namespaceBackplate {
   rect {
     fill: #ebeef0;
-    stroke: #b0b0b0;
-    stroke-dasharray: 12px 12px;
-    stroke-width: 5px;
+    stroke: #9ba1a8;
+    stroke-dasharray: 10px 10px;
+    stroke-width: 2px;
   }
 
   text {
     font-size: 1.75em;
-    fill: #7b7b7b;
+    fill: #58626d;
     font-weight: 700;
   }
 }
